Just a mini fix for modphp, because version NaN was pissing me off :)

+++ src/modules/modphp.cpp
@@ -31,9 +31,9 @@ extern "C" std::string name()
     return "ModPHP";
 }
 
-extern "C" int  version()
+extern "C" float version()
 {
-    return 5;
+    return 5.0;
 }
 
 ModPHP::ModPHP()
